requirements.txt
colorama==0.4.4
crayons==0.3.1
piexif==1.1.3
Pillow==6.2.2
pycryptodome==3.9.8
stegano==0.9.8
